remove admin home and update router/tests
